Remove View Logic from Route
The routes responsiblity is to pass along data and return a View,
not to figure out which View to return.  The responsiblity of which View to return is that of the View.
So here we add a Resource View component that
will contain the necessary logic to determine
which View to return to the Route.
