+++ src/components/resource/resource-view/index.js
@@ -1,0 +1,1 @@
+export { default } from './resource-view';
